+++ src/function/setENV.mjs
@@ -1,4 +1,4 @@
-import { Lodash as _, getStorage, log } from "@nsnanocat/util";
+import { Console, getStorage, Lodash as _ } from "@nsnanocat/util";
 
 /**
  * Set Environment Variables
@@ -9,12 +9,13 @@ import { Lodash as _, getStorage, log } from "@nsnanocat/util";
  * @return {Object} { Settings, Caches, Configs }
  */
 export default function setENV(name, platforms, database) {
-	log("☑️ Set Environment Variables", "");
+	Console.log("☑️ Set Environment Variables");
 	const { Settings, Caches, Configs } = getStorage(name, platforms, database);
 	/***************** Settings *****************/
-	log(`✅ Set Environment Variables, Settings: ${typeof Settings}, Settings内容: ${JSON.stringify(Settings)}`, "");
+	Console.debug(`typeof Settings: ${typeof Settings}`, `Settings: ${JSON.stringify(Settings)}`);
 	/***************** Caches *****************/
-	//log(`✅ Set Environment Variables, Caches: ${typeof Caches}, Caches内容: ${JSON.stringify(Caches)}`, "");
+	//Console.debug(`typeof Caches: ${typeof Caches}`, `Caches: ${JSON.stringify(Caches)}`);
 	/***************** Configs *****************/
+	Console.log("✅ Set Environment Variables");
 	return { Settings, Caches, Configs };
 };