+++ CHANGELOG.md
@@ -7,3 +7,4 @@
   * 升级了 `@nsnanocat/util`
     * `util` 由 `submodule` 更改为 `package`
     * `$platform` 改为 `$app`
+    * 使用了全新的 `Console` polyfill